Removed global variables, fixed headers, and fixed c files accordingly
Learned that extern variables are not the best ... so removed those and passed them through function calls. 
Also the old header (header.h) was incorrect. I had the wrong idea of headers. Created header files for each file (except for main).
Had to edit the .c files around a bit so the new implementation would work.